Module Smyth.Eval

type eval_result = (Lang.res * Lang.resumption_assertions, string) Stdlib.result

The return type of live evaluation and resumption. Live evaluation and resumption may fail (e.g., by running out of fuel), in which case they will return a string describing the error that occurred.

val eval : Lang.env -> Lang.exp -> eval_result

eval env exp live evaluates the expression exp in the environment env.

val resume : Lang.hole_filling -> Lang.res -> eval_result

resume hf res resumes the result res with the hole-filling hf.