Module Timer.Multi

A resettable countdown timer that accumulates time every time the accumulate function is called.

type t =

| Guess
The timer for raw term enumeration.

The available multi timers.

val reset : t -> unit

Resets the timer.

val accumulate : t -> (unit -> 'a) -> 'a

Accumulates additional time on the timer.

val check : t -> bool

Returns true if the timer is still running, and false if the elapsed time is greater than the cutoff.