Module Smyth__Branch

Type-and-example-directed branching.

This module implements type-and-example-directed branching as defined in Figure 8 of the ICFP 2020 paper.

val branch : int -> Smyth.Lang.hole_ctx -> Smyth.Lang.datatype_ctx -> Smyth.Lang.hole_filling -> Smyth.Lang.synthesis_goal -> ((Smyth.Lang.exp * Smyth.Lang.fill_goal list) * Smyth.Lang.constraints) Smyth.Nondet.t

branch max_scrutinee_size delta sigma hf goal splits the a synthesisi goal goal into subgoals based on type-and-example-directed branching with a maximum synthesized scrutinee size of max_scrutinee_size.