Module Smyth__Constraints

Hole fillings, unfilled holes (unsolved constraints), and unevaluation constraints.

This module provides functions that operate on the various constraint data structures used in Smyth.

val delete_min : 'a Smyth.Lang.hole_map -> ((Smyth.Lang.hole_name * 'a) * 'a Smyth.Lang.hole_map) option

Deletes and returns the minimum element of a hole map (such as a Lang.unsolved_constraints).

val empty : Smyth.Lang.constraints

The empty unevaluation constraint.

val from_hole_filling : Smyth.Lang.hole_filling -> Smyth.Lang.constraints

from_hole_filling hf creates a new unevaluation constraint consisting of the hole filling hf and no unfilled holes.

val from_unsolved_constraints : Smyth.Lang.unsolved_constraints -> Smyth.Lang.constraints

from_unsolved_constraints us creates a new unevaluation constraint consisting of the unfilled holes us and no hole filling.

val solved_singleton : Smyth.Lang.hole_name -> Smyth.Lang.exp -> Smyth.Lang.constraints

solved_singleton h e creates a new unevaluation constraint consisting of the singleton hole filling h -> e and no unfilled holes.

val unsolved_singleton : Smyth.Lang.hole_name -> Smyth.Lang.worlds -> Smyth.Lang.constraints

solved_singleton h x creates a new unevaluation constraint consisting of the singleton unfilled hole h -> x and no hole filling.

val merge_solved : Smyth.Lang.hole_filling list -> Smyth.Lang.hole_filling option

Syntactically merges a list of hole fillings, failing if a hole is bound to two syntactically unequal expressions.

val merge_unsolved : Smyth.Lang.unsolved_constraints list -> Smyth.Lang.unsolved_constraints

Syntactically merges a list of unfilled holes.

val merge : Smyth.Lang.constraints list -> Smyth.Lang.constraints option

Syntactically merges a list of unevaluation constraints using merge_solved and merge_unsolved.

val satisfies : Smyth.Lang.hole_filling -> Smyth.Lang.constraints -> bool

Unevaluation constraint satisfaction as defined by Figure 5 of the ICFP 2020 paper.

satisfies hf ks checks if a hole filling hf satisfies a set of uenvaluation constraints ks using Example.exp_satisfies (i.e., the notion of "ground truth" satisfaction).