Module Smyth__Env

Evaluation environments.

This module provides functions that operate on the evaluation environment data structure used in Smyth.

Entries closer to the "beginning" of the environment shadow entries appearing later. Entries are typically "consed" to the beginning of an environment (for example, with add_res).

val empty : Smyth.Lang.env

The empty evaluation environment.

val all_res : Smyth.Lang.env -> (string * Smyth.Lang.res) list

all_res env returns all name-to-Lang.res bindings in the environment env; that is, the assignment of expression variables to results.

val all_type : Smyth.Lang.env -> (string * Smyth.Lang.typ) list

all_type env returns all name-to-Lang.typ bindings in the environment env; that is, the assignment of type variables to types.

val concat : Smyth.Lang.env list -> Smyth.Lang.env

Concatenates a list of environments.

val add_res : (string * Smyth.Lang.res) -> Smyth.Lang.env -> Smyth.Lang.env

Adds a name-to-Lang.res binding to the beginning of an environment.

val concat_res : (string * Smyth.Lang.res) list -> Smyth.Lang.env -> Smyth.Lang.env

Concatenates a list of name-to-Lang.res bindings to the beginning of an environment.

val add_type : (string * Smyth.Lang.typ) -> Smyth.Lang.env -> Smyth.Lang.env

Adds a name-to-Lang.typ binding to the beginning of an environment.

val concat_type : (string * Smyth.Lang.typ) list -> Smyth.Lang.env -> Smyth.Lang.env

Concatenates a list of name-to-Lang.typ bindings to the beginning of an environment.