Module Smyth__Eval

Live evaluation and resumption.

This module implements live evaluation and resumption as defined in Figure 4 of the ICFP 2020 paper. Figure 7 also defines "program evaluation"; that concept is simply included in the definition of live evaluation here.

type eval_result = (Smyth.Lang.res * Smyth.Lang.resumption_assertions, string) Stdlib.result

The return type of live evaluation and resumption. Live evaluation and resumption may fail (e.g., by running out of fuel), in which case they will return a string describing the error that occurred.

val eval : Smyth.Lang.env -> Smyth.Lang.exp -> eval_result

eval env exp live evaluates the expression exp in the environment env.

val resume : Smyth.Lang.hole_filling -> Smyth.Lang.res -> eval_result

resume hf res resumes the result res with the hole-filling hf.