Module Smyth__Fill

Type-and-example-directed hole synthesis.

This module implements type-and-example-directed branching as defined in Figure 8 of the ICFP 2020 paper.

val fill : Smyth.Lang.synthesis_params -> Smyth.Lang.hole_ctx -> Smyth.Lang.datatype_ctx -> Smyth.Lang.hole_filling -> Smyth.Lang.fill_goal -> (Smyth.Lang.constraints * Smyth.Lang.hole_ctx) Smyth.Nondet.t

fill params delta sigma f g performs type-and-example-directed hole synthesis over the goal defined by g as defined by Figure 8 of the ICFP 2020 paper. Relies on Refine and Branch for the rule Refine-or-Branch and Term_gen for the rule Guess-and-Check.