Module Smyth__Params

Synthesis and system parameters.

Many of these options are also configurable with command-line flags.

Note: all timing units are in seconds.

val version : string

The overall version of the system.

val debug_mode : bool Stdlib.ref

Whether or not debugging should be enabled.

type ranking_method =

| Size

The available methods for ranking synthesis solutions.

val ranking_method : ranking_method Stdlib.ref

The method to rank synthesis solutions.

val max_solution_count : int option Stdlib.ref

The maximum synthesis solution count. None indicates no bound/infinity.

val uneval_case_budget : int Stdlib.ref

The unevaluation case budget; sets a limit to the number of times U-Case is allowed to be called.

val uneval_limiter : int Stdlib.ref

The unevluation limiter; immediately cuts off any unevaluation path that exceeds a certain number of nondeterministic possibilities (thereby eliminating pathological search paths that are unlikely to succeed).

val max_total_time : float Stdlib.ref

The total time a single synthesis task is allowed to use.

val max_eval_time : float Stdlib.ref

The time a single call to evaluation is allowed to use.

val max_guess_time : float Stdlib.ref

The time a single call to Guess is allowed to use.

val initial_fuel : int Stdlib.ref

The initial fuel count for evaluation and resumption.

val log_info : bool Stdlib.ref

Whether or not to enable info-level logging.

val log_warn : bool Stdlib.ref

Whether or not to enable warn-level logging.