Module Smyth__Timer.Single

A countdown timer that, once started, continues to completion.

type t =

| Total
The timer for the total synthesis time.

| Eval
The timer for live evaluation and resumption.

The available single timers.

val start : t -> unit

Starts the timer.

val elapsed : t -> float

Returns how much time has elapsed since starting the timer.

val check : t -> bool

Returns true if the timer is still running, and false if the elapsed time is greater than the cutoff.