Module Smyth__Type

Type helpers and type-checking.

Helpers

val equal : Smyth.Lang.typ -> Smyth.Lang.typ -> bool

equal tau1 tau2 determines if tau1 and tau2 have the same abstract syntax tree.

val wildcard : Smyth.Lang.typ

The "wildcard" type.

val matches : Smyth.Lang.typ -> Smyth.Lang.typ -> bool

matches tau1 tau2 determines if tau1 and tau2 have the same abstract syntax tree modulo any wildcards.

val is_base : Smyth.Lang.typ -> bool

is_base tau returns determines if tau is a base type.

val domain_of_codomain : codomain:Smyth.Lang.typ -> Smyth.Lang.typ -> Smyth.Lang.typ option

domain_of_codomain ~codomain root looks for codomain on the right-hand side of a type arrow in root that is not on the left-hand side of any type arrow and, if found, returns the left-hand side of that type arrow.

For example, suppose tau = a -> (b -> (c -> d)). Then:

domain_of_codomain ~codomain:d tau = Some c
domain_of_codomain ~codomain:(c -> d) tau = Some b
domain_of_codomain ~codomain:(b -> (c -> d)) tau = Some a
domain_of_codomain ~codomain:(a -> (b -> (c -> d))) tau = None

Also:

domain_of_codomain ~codomain:a a = None
domain_of_codomain ~codomain:a (() -> a) = Some ()

val bind_spec : Smyth.Lang.type_ctx -> Smyth.Lang.exp -> Smyth.Lang.bind_spec

bind_spec gamma e returns the computed binding specification of e with respect to gamma. In particular:

Projections decrease the binding specification of their argument

Variables are looked up in gamma

val sub_bind_spec : Smyth.Lang.bind_spec -> Smyth.Lang.bind_spec

sub_bind_spec b "decreases" the binding specification b. In particular, if b is a binding specification for the argument of some function, then sub_bind_spec b will be a binding specification for decreasing on that function.

val structurally_decreasing_bind_spec : head_spec:Smyth.Lang.bind_spec -> arg_spec:Smyth.Lang.bind_spec -> bool

structurally_decreasing_bind_spec ~head_spec ~arg_spec determines if arg_spec is decreasing on head_spec. If head_spec is not a binding specification for a recursive function, then structurally_decreasing ~head_spec ~arg_spec will return true.

val structurally_decreasing : Smyth.Lang.type_ctx -> head:Smyth.Lang.exp -> arg:Smyth.Lang.exp -> bool

structurally_decreasing gamma ~head ~arg determines if arg is structurally decreasing on head in the context gamma.

val matches_dec : string option -> Smyth.Lang.bind_spec -> bool

matches_dec (Some f) bind_spec determines if bind_spec is structurally decreasing on the recursive function f. matches_dec None bind_spec will always return true.

val peel_forall : Smyth.Lang.typ -> string list * Smyth.Lang.typ

Peels all outer universal quantifiers on a type into a list. For example:

peel_forall (forall a . forall b . a * b) = (["a", "b"], a * b)
peel_forall a = ([], a)
peel_forall (int * int) = ([], int * int)
peel_forall (forall c . int) = (["c"], int)
peel_forall (forall c . d) = (["c"], d)

Substitution

val substitute : before:string -> after:Smyth.Lang.typ -> Smyth.Lang.typ -> Smyth.Lang.typ

substitute ~before ~after tau substitutes replaces free occurrences of before in tau with after.

val substitute_many : bindings:(string * Smyth.Lang.typ) list -> Smyth.Lang.typ -> Smyth.Lang.typ

Performs many substitutions one after another; see substitute.

Type-checking

The following functions implement a bidirectional type checker.

type error =

| VarNotFound of string

| CtorNotFound of string

| PatternMatchFailure of Smyth.Lang.typ * Smyth.Lang.pat

| WrongNumberOfTypeArguments of int * int

| GotFunctionButExpected of Smyth.Lang.typ

| GotTupleButExpected of Smyth.Lang.typ

| GotTypeAbstractionButExpected of Smyth.Lang.typ

| GotButExpected of Smyth.Lang.typ * Smyth.Lang.typ

| BranchMismatch of string * string

| CannotInferFunctionType

| CannotInferCaseType

| CannotInferHoleType

| ExpectedArrowButGot of Smyth.Lang.typ

| ExpectedTupleButGot of Smyth.Lang.typ

| ExpectedForallButGot of Smyth.Lang.typ

| ExpectedDatatypeButGot of Smyth.Lang.typ

| TupleLengthMismatch of Smyth.Lang.typ

| ProjectionLengthMismatch of Smyth.Lang.typ

| ProjectionOutOfBounds of int * int

| TypeAbstractionParameterNameMismatch of string * string

| AssertionTypeMismatch of Smyth.Lang.typ * Smyth.Lang.typ

Type-checking errors.

val check : Smyth.Lang.datatype_ctx -> Smyth.Lang.type_ctx -> Smyth.Lang.exp -> Smyth.Lang.typ -> (Smyth.Lang.hole_ctx, Smyth.Lang.exp * error) Stdlib.result

Bidirectional type "checking".

val infer : Smyth.Lang.datatype_ctx -> Smyth.Lang.type_ctx -> Smyth.Lang.exp -> (Smyth.Lang.typ * Smyth.Lang.hole_ctx, Smyth.Lang.exp * error) Stdlib.result

Bidirectional type "inference" or "synthesis".