Module Smyth__Uneval

Live bidirectional example checking via live unevaluation, as defined in Figure 6 of the ICFP 2020 paper.

val minimal_uneval : bool Stdlib.ref

Whether or not unevaluation should proceed in the "minimal" mode. Minimal mode uses U-Case-Guess and not U-Case; non-minimal mode does the opposite.

val check : Smyth.Lang.hole_ctx -> Smyth.Lang.datatype_ctx -> Smyth.Lang.hole_filling -> Smyth.Lang.exp -> Smyth.Lang.worlds -> Smyth.Lang.constraints Smyth.Nondet.t

Live bidirectional example checking.

val uneval : Smyth.Lang.hole_ctx -> Smyth.Lang.datatype_ctx -> Smyth.Lang.hole_filling -> Smyth.Lang.res -> Smyth.Lang.example -> Smyth.Lang.constraints Smyth.Nondet.t

Live unevaluation.

val simplify_assertions : Smyth.Lang.hole_ctx -> Smyth.Lang.datatype_ctx -> Smyth.Lang.resumption_assertions -> Smyth.Lang.constraints Smyth.Nondet.t

Assertion simplification, as defined in Figure 7 of the ICFP 2020 paper.